requirements.txt
discord.py>=2.2.2
requests>=2.28.0
python-dotenv>=1.0.0
pillow>=9.5.0